Add files via upload
This research thesis is focused on developing optimal allocation of Uber drivers to
the most profitable and economically feasible geohash locations in a city. In doing that,
we aspire to provide a win-win solution for all the stakeholders including passenger,
driver as well as the ride sharing company like Uber,Lyft etc. We start by dividing
the city into equally spaced clusters and then calculating the distance between drivers,
passengers and final destination of passengers. Subsequently, we calculate expected
profits for drivers considering their current location, location of customer request and
the travel destination of passenger. Through this optimal allocation model, we propose
to reduce driver’s idle cruising time and the waiting time of passenger after requesting
a ride. Thus, the current imbalance between supply and demand of Uber drivers is
expected to be minimized